Replication Service
A sophisticated Replication Service built into the Directory Service.

When a file service is detected to have died, the replication service
invalidates all the failed services files, redirects those requests to
it’s backup file service and then prepares the files for replication
into their new backup service.

A continuously running task for replicating the files runs in the
background of the application and is called once every 30 seconds.