Lock Service and File Updating
A full lock service that allows users to securely and confidently
update their files knowing their upload will not get overwritten.

Any user with access to a file can now lock, unlock and update a file.
The only criteria is that only one lock can be distributed for any file
at a time and that the updated file is of the same type and name as the
original.